+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.21.1
+
+### Fixed
+
+- **Default Values Can Be Specified with allOf and anyOf Schemas**
+
+  Schemas that use the `allOf` or `anyOf` property can now specify a default value at the shared schema level. For example:
+
+  ```yaml
+  MySchema:
+    default: false
+    anyOf:
+      - type: boolean
+      - type: number
+  ```
+
 ## 1.21.0
 
 ### Added